+++ asymmetric.py
